requirements.txt
aioredis==2.0.1
alembic==1.13.2
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
bcrypt==4.1.2
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.1
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.108.0
fastapi-users==13.0.0
fastapi-users-db-sqlalchemy==6.0.1
flower==2.0.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
humanize==4.10.0
idna==3.6
Jinja2==3.1.4
kombu==5.4.2
makefun==1.15.4
Mako==1.3.5
MarkupSafe==2.1.5
outcome==1.3.0.post0
packaging==24.1
prometheus_client==0.21.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pwdlib==0.2.0
pycparser==2.22
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.1
redis==5.0.8
requests==2.32.3
selenium==4.25.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.25
SQLAlchemy-Utils==0.41.1
starlette==0.32.0.post1
tornado==6.4.1
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.25.0
uvloop==0.19.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==12.0
wsproto==1.2.0
